feat: 4-directional cursor movement. nuance: "memoryless". see details...
The cursor in vim and VSCode has a "memory" and true underlying value.
For example, you are at column 10 on row 0. You move down to row 1
but there is no characters or less characters than the contents on row 0.
The cursor is positioned at the farthest right character. In the special
case where there is no characters, the cursor will occupy col 0 even though
there isn't a character there. This is why we have to do Math.max(0, len(str on line) -1)
throughout the code.

At any rate, if you move up, you will get back to column 10 in vim/vscode.
In our implementation, we just go to column 0 in row 0. if we were on column 3
of row 1 then we'd go to column 3 of row 0.